Feature/getshelterbyid

+++ Lapka.Identity.Application/Commands/Handlers/DeleteShelterHandler.cs
@@ -1,6 +1,9 @@
 ﻿using System.Threading.Tasks;
 using Convey.CQRS.Commands;
+using Lapka.Identity.Application.Dto;
+using Lapka.Identity.Application.Exceptions;
 using Lapka.Identity.Application.Services;
+using Lapka.Identity.Core.Entities;
 
 namespace Lapka.Identity.Application.Commands.Handlers
 {
@@ -16,7 +19,9 @@ namespace Lapka.Identity.Application.Commands.Handlers
         }
         public async Task HandleAsync(DeleteShelter command)
         {
-            var shelter = await _shelterRepository.GetByIdAsync(command.Id);
+            Shelter shelter = await _shelterRepository.GetByIdAsync(command.Id);
+            if (shelter is null) throw new ShelterNotFoundException();
+            
             shelter.Delete();
             
             await _shelterRepository.DeleteAsync(shelter);
